Refactor: separate tests by function
테스트를 기능별로 분리함
- 이 과정에서, FixedLadderCreator가 부가적으로 생성됨
