+++ src/main/java/ladder/game/LadderGameFactory.java
@@ -1,13 +1,20 @@
 package ladder.game;
 
+import ladder.creator.FixedLadderCreator;
 import ladder.wrapper.NaturalNumber;
 import ladder.creator.LadderCreator;
 import ladder.creator.RandomLadderCreator;
 
 public class LadderGameFactory {
     public static LadderGame createRandomLadderGame(NaturalNumber numberOfRows, NaturalNumber numberOfPerson) {
-        LadderCreator ladderRandomCreator = new RandomLadderCreator();
+        LadderCreator randomLadderCreator = new RandomLadderCreator();
 
-        return new LadderGame(ladderRandomCreator, numberOfRows, numberOfPerson);
+        return new LadderGame(randomLadderCreator, numberOfRows, numberOfPerson);
+    }
+
+    public static LadderGame createFixedLadderGame(NaturalNumber numberOfRows, NaturalNumber numberOfPerson) {
+        LadderCreator fixedLadderCreator = new FixedLadderCreator();
+
+        return new LadderGame(fixedLadderCreator, numberOfRows, numberOfPerson);
     }
 }
